fixed decimal

+++ makemkv/makemkv.py
@@ -243,7 +243,7 @@ class MakeMKV:
         elif key == "framerate":
             # convert "##.### (#####/####)" to int string
             if m := re.match(r"^(\d+(?:\.\d+)*)\s\(\d+/\d+\)$", value):
-                return_value = float(m[1])
+                return_value = Decimal(m[1])
             else:
                 return_value = int(value)
         elif key == "segments_map":
